refactor: use order class in contracts-ts
