+++ packages/contracts-ts/src/order.ts
@@ -1,4 +1,4 @@
-import { AdapterTypes } from '@cowprotocol/common'
+import { AbstractProviderAdapter, AdapterTypes } from '@cowprotocol/common'
 
 /**
  * Gnosis Protocol v2 order data.
@@ -229,4 +229,131 @@ export interface OrderUidParams {
    * The timestamp this order is valid until.
    */
   validTo: number | Date
+}
+
+export class ContractsTs_Order<T extends AdapterTypes = AdapterTypes> {
+  constructor(private adapter: AbstractProviderAdapter<T>) {
+    this.adapter = adapter
+  }
+
+  /**
+   * Normalizes a timestamp value to a Unix timestamp.
+   * @param time The timestamp value to normalize.
+   * @return Unix timestamp or number of seconds since the Unix Epoch.
+   */
+  public timestamp(t: Timestamp): number {
+    return timestamp(t)
+  }
+
+  /**
+   * Normalizes an app data value to a 32-byte hash.
+   * @param hashLike A hash-like value to normalize.
+   * @returns A 32-byte hash encoded as a hex-string.
+   */
+  public hashify<T extends AdapterTypes>(h: HashLike<T>): string {
+    return typeof h === 'number' ? `0x${h.toString(16).padStart(64, '0')}` : this.adapter.hexZeroPad(h, 32)
+  }
+
+  /**
+   * Normalizes the balance configuration for a buy token. Specifically, this
+   * function ensures that {@link OrderBalance.EXTERNAL} gets normalized to
+   * {@link OrderBalance.ERC20}.
+   *
+   * @param balance The balance configuration.
+   * @returns The normalized balance configuration.
+   */
+  public normalizeBuyTokenBalance(balance: OrderBalance | undefined): OrderBalance.ERC20 | OrderBalance.INTERNAL {
+    return normalizeBuyTokenBalance(balance)
+  }
+
+  /**
+   * Normalizes an order for hashing and signing, so that it can be used with
+   * Ethers.js for EIP-712 operations.
+   * @param hashLike A hash-like value to normalize.
+   * @returns A 32-byte hash encoded as a hex-string.
+   */
+  public normalizeOrder(order: Order<T>): NormalizedOrder<T> {
+    if (order.receiver === this.adapter.ZERO_ADDRESS) {
+      throw new Error('receiver cannot be address(0)')
+    }
+
+    const normalizedOrder = {
+      ...order,
+      sellTokenBalance: order.sellTokenBalance ?? OrderBalance.ERC20,
+      receiver: order.receiver ?? this.adapter.ZERO_ADDRESS,
+      validTo: timestamp(order.validTo),
+      appData: this.hashify(order.appData),
+      buyTokenBalance: normalizeBuyTokenBalance(order.buyTokenBalance),
+    }
+    return normalizedOrder
+  }
+
+  /**
+   * Compute the 32-byte signing hash for the specified order.
+   *
+   * @param domain The EIP-712 domain separator to compute the hash for.
+   * @param types The order to compute the digest for.
+   * @return Hex-encoded 32-byte order digest.
+   */
+  public hashTypedData(
+    domain: T['TypedDataDomain'],
+    types: T['TypedDataTypes'],
+    data: Record<string, unknown>,
+  ): string {
+    return this.adapter.hashTypedData(domain, types, data)
+  }
+
+  /**
+   * Compute the 32-byte signing hash for the specified order.
+   *
+   * @param domain The EIP-712 domain separator to compute the hash for.
+   * @param order The order to compute the digest for.
+   * @return Hex-encoded 32-byte order digest.
+   */
+  public hashOrder(domain: T['TypedDataDomain'], order: Order<T>): string {
+    return this.hashTypedData(domain, { Order: ORDER_TYPE_FIELDS }, this.normalizeOrder(order))
+  }
+
+  /**
+   * Computes the order UID for an order and the given owner.
+   */
+  public computeOrderUid(domain: T['TypedDataDomain'], order: Order<T>, owner: string): string {
+    return this.packOrderUidParams({
+      orderDigest: this.hashOrder(domain, order),
+      owner,
+      validTo: order.validTo,
+    })
+  }
+
+  /**
+   * Compute the unique identifier describing a user order in the settlement
+   * contract.
+   *
+   * @param OrderUidParams The parameters used for computing the order's unique
+   * identifier.
+   * @returns A string that unequivocally identifies the order of the user.
+   */
+  public packOrderUidParams({ orderDigest, owner, validTo }: OrderUidParams): string {
+    return this.adapter.solidityPack(['bytes32', 'address', 'uint32'], [orderDigest, owner, timestamp(validTo)])
+  }
+
+  /**
+   * Extracts the order unique identifier parameters from the specified bytes.
+   *
+   * @param orderUid The order UID encoded as a hexadecimal string.
+   * @returns The extracted order UID parameters.
+   */
+  public extractOrderUidParams(orderUid: string): OrderUidParams {
+    const bytes = this.adapter.arrayify(orderUid)
+    if (bytes.length != ORDER_UID_LENGTH) {
+      throw new Error('invalid order UID length')
+    }
+
+    const view = new DataView(bytes.buffer)
+    return {
+      orderDigest: this.adapter.hexlify(bytes.subarray(0, 32)),
+      owner: this.adapter.getChecksumAddress(this.adapter.hexlify(bytes.subarray(32, 52))),
+      validTo: view.getUint32(52),
+    }
+  }
 }
